requirement.txt
Flask==3.0.0
opencv-python==4.10.0.84
tensorflow==2.19.0
mtcnn==0.1.1
xgboost==2.0.0
scikit-learn==1.3.0
joblib==1.3.1
numpy==1.24.3
Pillow==10.0.0